:recycle: refactor: 에러 처리 구조 개선 (#11) - 에러 코드 패키지 구조 개선 (GlobalErrorCode, ErrorCode) - 서비스 레이어 호환성 수정 - 관련 서비스 코드 리팩토링
