+++ src/main/java/com/adit/backend/domain/place/service/CommonPlaceService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.adit.backend.domain.place.dto.request.CommonPlaceRequestDto;
 import com.adit.backend.domain.place.entity.CommonPlace;
 import com.adit.backend.domain.place.repository.CommonPlaceRepository;
+import com.adit.backend.global.error.GlobalErrorCode;
 import com.adit.backend.global.error.exception.BusinessException;
-import com.adit.backend.global.error.exception.GlobalErrorCode;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
